refactor test to match main directory; disable tests
