+++ src/test/java/seedu/zerotoone/logic/commands/exercise/DeleteCommandTest.java
@@ -1,0 +1,109 @@
+package seedu.zerotoone.logic.commands.exercise;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+// import static seedu.zerotoone.logic.commands.CommandTestUtil.assertCommandFailure;
+// import static seedu.zerotoone.logic.commands.CommandTestUtil.assertCommandSuccess;
+// import static seedu.zerotoone.logic.commands.CommandTestUtil.showExerciseAtIndex;
+// import static seedu.zerotoone.testutil.TypicalExercises.getTypicalExerciseList;
+import static seedu.zerotoone.testutil.TypicalIndexes.INDEX_FIRST_EXERCISE;
+import static seedu.zerotoone.testutil.TypicalIndexes.INDEX_SECOND_EXERCISE;
+
+import org.junit.jupiter.api.Test;
+
+import seedu.zerotoone.commons.core.Messages;
+import seedu.zerotoone.commons.core.index.Index;
+import seedu.zerotoone.model.Model;
+import seedu.zerotoone.model.ModelManager;
+import seedu.zerotoone.model.exercise.Exercise;
+import seedu.zerotoone.model.userprefs.UserPrefs;
+
+/**
+ * Contains integration tests (interaction with the Model, UndoCommand and RedoCommand) and unit tests for
+ * {@code DeleteCommand}.
+ */
+public class DeleteCommandTest {
+
+    // private Model model = new ModelManager(getTypicalExerciseList(), new UserPrefs());
+
+    // @Test
+    // public void execute_validIndexUnfilteredList_success() {
+    //     Exercise exerciseToDelete = model.getFilteredExerciseList().get(INDEX_FIRST_EXERCISE.getZeroBased());
+    //     DeleteCommand deleteCommand = new DeleteCommand(INDEX_FIRST_EXERCISE);
+
+    //     String expectedMessage = String.format(DeleteCommand.MESSAGE_DELETE_EXERCISE_SUCCESS, exerciseToDelete);
+
+    //     ModelManager expectedModel = new ModelManager(model.getExerciseList(), new UserPrefs());
+    //     expectedModel.deleteExercise(exerciseToDelete);
+
+    //     assertCommandSuccess(deleteCommand, model, expectedMessage, expectedModel);
+    // }
+
+    // @Test
+    // public void execute_invalidIndexUnfilteredList_throwsCommandException() {
+    //     Index outOfBoundIndex = Index.fromOneBased(model.getFilteredExerciseList().size() + 1);
+    //     DeleteCommand deleteCommand = new DeleteCommand(outOfBoundIndex);
+
+    //     assertCommandFailure(deleteCommand, model, Messages.MESSAGE_INVALID_EXERCISE_DISPLAYED_INDEX);
+    // }
+
+    // @Test
+    // public void execute_validIndexFilteredList_success() {
+    //     showExerciseAtIndex(model, INDEX_FIRST_EXERCISE);
+
+    //     Exercise exerciseToDelete = model.getFilteredExerciseList().get(INDEX_FIRST_EXERCISE.getZeroBased());
+    //     DeleteCommand deleteCommand = new DeleteCommand(INDEX_FIRST_EXERCISE);
+
+    //     String expectedMessage = String.format(DeleteCommand.MESSAGE_DELETE_EXERCISE_SUCCESS, exerciseToDelete);
+
+    //     Model expectedModel = new ModelManager(model.getExerciseList(), new UserPrefs());
+    //     expectedModel.deleteExercise(exerciseToDelete);
+    //     showNoExercise(expectedModel);
+
+    //     assertCommandSuccess(deleteCommand, model, expectedMessage, expectedModel);
+    // }
+
+    // @Test
+    // public void execute_invalidIndexFilteredList_throwsCommandException() {
+    //     showExerciseAtIndex(model, INDEX_FIRST_EXERCISE);
+
+    //     Index outOfBoundIndex = INDEX_SECOND_EXERCISE;
+    //     // ensures that outOfBoundIndex is still in bounds of address book list
+    //     assertTrue(outOfBoundIndex.getZeroBased() < model.getExerciseList().getExerciseList().size());
+
+    //     DeleteCommand deleteCommand = new DeleteCommand(outOfBoundIndex);
+
+    //     assertCommandFailure(deleteCommand, model, Messages.MESSAGE_INVALID_EXERCISE_DISPLAYED_INDEX);
+    // }
+
+    // @Test
+    // public void equals() {
+    //     DeleteCommand deleteFirstCommand = new DeleteCommand(INDEX_FIRST_EXERCISE);
+    //     DeleteCommand deleteSecondCommand = new DeleteCommand(INDEX_SECOND_EXERCISE);
+
+    //     // same object -> returns true
+    //     assertTrue(deleteFirstCommand.equals(deleteFirstCommand));
+
+    //     // same values -> returns true
+    //     DeleteCommand deleteFirstCommandCopy = new DeleteCommand(INDEX_FIRST_EXERCISE);
+    //     assertTrue(deleteFirstCommand.equals(deleteFirstCommandCopy));
+
+    //     // different types -> returns false
+    //     assertFalse(deleteFirstCommand.equals(1));
+
+    //     // null -> returns false
+    //     assertFalse(deleteFirstCommand.equals(null));
+
+    //     // different exercise -> returns false
+    //     assertFalse(deleteFirstCommand.equals(deleteSecondCommand));
+    // }
+
+    // /**
+    //  * Updates {@code model}'s filtered list to show no one.
+    //  */
+    // private void showNoExercise(Model model) {
+    //     model.updateFilteredExerciseList(p -> false);
+
+    //     assertTrue(model.getFilteredExerciseList().isEmpty());
+    // }
+}
