--- src/test/java/seedu/zerotoone/logic/parser/FindCommandParserTest.java
@@ -1,34 +1,0 @@
-package seedu.zerotoone.logic.parser;
-
-import static seedu.zerotoone.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.zerotoone.logic.parser.CommandParserTestUtil.assertParseFailure;
-import static seedu.zerotoone.logic.parser.CommandParserTestUtil.assertParseSuccess;
-
-import java.util.Arrays;
-
-import org.junit.jupiter.api.Test;
-
-import seedu.zerotoone.logic.commands.FindCommand;
-import seedu.zerotoone.model.exercise.NameContainsKeywordsPredicate;
-
-public class FindCommandParserTest {
-
-    private FindCommandParser parser = new FindCommandParser();
-
-    @Test
-    public void parse_emptyArg_throwsParseException() {
-        assertParseFailure(parser, "     ", String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
-    }
-
-    @Test
-    public void parse_validArgs_returnsFindCommand() {
-        // no leading and trailing whitespaces
-        FindCommand expectedFindCommand =
-                new FindCommand(new NameContainsKeywordsPredicate(Arrays.asList("Alice", "Bob")));
-        assertParseSuccess(parser, "Alice Bob", expectedFindCommand);
-
-        // multiple whitespaces between keywords
-        assertParseSuccess(parser, " \n Alice \n \t Bob  \t", expectedFindCommand);
-    }
-
-}
