+++ src/test/java/seedu/zerotoone/model/userprefs/UserPrefsTest.java
@@ -1,0 +1,23 @@
+package seedu.zerotoone.model.userprefs;
+
+import static seedu.zerotoone.testutil.Assert.assertThrows;
+
+import org.junit.jupiter.api.Test;
+
+import seedu.zerotoone.model.userprefs.UserPrefs;
+
+public class UserPrefsTest {
+
+    // @Test
+    // public void setGuiSettings_nullGuiSettings_throwsNullPointerException() {
+    //     UserPrefs userPref = new UserPrefs();
+    //     assertThrows(NullPointerException.class, () -> userPref.setGuiSettings(null));
+    // }
+
+    // @Test
+    // public void setExerciseListFilePath_nullPath_throwsNullPointerException() {
+    //     UserPrefs userPrefs = new UserPrefs();
+    //     assertThrows(NullPointerException.class, () -> userPrefs.setExerciseListFilePath(null));
+    // }
+
+}
