+++ src/test/java/seedu/zerotoone/testutil/EditExerciseDescriptorBuilder.java
@@ -4,85 +4,85 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import seedu.zerotoone.logic.commands.EditCommand;
-import seedu.zerotoone.logic.commands.EditCommand.EditExerciseDescriptor;
-import seedu.zerotoone.model.exercise.Address;
-import seedu.zerotoone.model.exercise.Email;
-import seedu.zerotoone.model.exercise.Exercise;
-import seedu.zerotoone.model.exercise.Name;
-import seedu.zerotoone.model.exercise.Phone;
-import seedu.zerotoone.model.tag.Tag;
+// import seedu.zerotoone.logic.commands.EditCommand;
+// import seedu.zerotoone.logic.commands.EditCommand.EditExerciseDescriptor;
+// import seedu.zerotoone.model.exercise.Address;
+// import seedu.zerotoone.model.exercise.Email;
+// import seedu.zerotoone.model.exercise.Exercise;
+// import seedu.zerotoone.model.exercise.Name;
+// import seedu.zerotoone.model.exercise.Phone;
+// import seedu.zerotoone.model.tag.Tag;
 
 /**
  * A utility class to help with building EditExerciseDescriptor objects.
  */
 public class EditExerciseDescriptorBuilder {
 
-    private EditExerciseDescriptor descriptor;
+    // private EditExerciseDescriptor descriptor;
 
-    public EditExerciseDescriptorBuilder() {
-        descriptor = new EditExerciseDescriptor();
-    }
+    // public EditExerciseDescriptorBuilder() {
+    //     descriptor = new EditExerciseDescriptor();
+    // }
 
-    public EditExerciseDescriptorBuilder(EditExerciseDescriptor descriptor) {
-        this.descriptor = new EditExerciseDescriptor(descriptor);
-    }
+    // public EditExerciseDescriptorBuilder(EditExerciseDescriptor descriptor) {
+    //     this.descriptor = new EditExerciseDescriptor(descriptor);
+    // }
 
-    /**
-     * Returns an {@code EditExerciseDescriptor} with fields containing {@code exercise}'s details
-     */
-    public EditExerciseDescriptorBuilder(Exercise exercise) {
-        descriptor = new EditCommand.EditExerciseDescriptor();
-        descriptor.setName(exercise.getName());
-        // descriptor.setPhone(exercise.getPhone());
-        // descriptor.setEmail(exercise.getEmail());
-        // descriptor.setAddress(exercise.getAddress());
-        // descriptor.setTags(exercise.getTags());
-    }
+    // /**
+    //  * Returns an {@code EditExerciseDescriptor} with fields containing {@code exercise}'s details
+    //  */
+    // public EditExerciseDescriptorBuilder(Exercise exercise) {
+    //     descriptor = new EditCommand.EditExerciseDescriptor();
+    //     descriptor.setName(exercise.getName());
+    //     // descriptor.setPhone(exercise.getPhone());
+    //     // descriptor.setEmail(exercise.getEmail());
+    //     // descriptor.setAddress(exercise.getAddress());
+    //     // descriptor.setTags(exercise.getTags());
+    // }
 
-    /**
-     * Sets the {@code Name} of the {@code EditExerciseDescriptor} that we are building.
-     */
-    public EditExerciseDescriptorBuilder withName(String name) {
-        descriptor.setName(new Name(name));
-        return this;
-    }
+    // /**
+    //  * Sets the {@code Name} of the {@code EditExerciseDescriptor} that we are building.
+    //  */
+    // public EditExerciseDescriptorBuilder withName(String name) {
+    //     descriptor.setName(new Name(name));
+    //     return this;
+    // }
 
-    /**
-     * Sets the {@code Phone} of the {@code EditExerciseDescriptor} that we are building.
-     */
-    public EditExerciseDescriptorBuilder withPhone(String phone) {
-        descriptor.setPhone(new Phone(phone));
-        return this;
-    }
+    // /**
+    //  * Sets the {@code Phone} of the {@code EditExerciseDescriptor} that we are building.
+    //  */
+    // public EditExerciseDescriptorBuilder withPhone(String phone) {
+    //     descriptor.setPhone(new Phone(phone));
+    //     return this;
+    // }
 
-    /**
-     * Sets the {@code Email} of the {@code EditExerciseDescriptor} that we are building.
-     */
-    public EditExerciseDescriptorBuilder withEmail(String email) {
-        descriptor.setEmail(new Email(email));
-        return this;
-    }
+    // /**
+    //  * Sets the {@code Email} of the {@code EditExerciseDescriptor} that we are building.
+    //  */
+    // public EditExerciseDescriptorBuilder withEmail(String email) {
+    //     descriptor.setEmail(new Email(email));
+    //     return this;
+    // }
 
-    /**
-     * Sets the {@code Address} of the {@code EditExerciseDescriptor} that we are building.
-     */
-    public EditExerciseDescriptorBuilder withAddress(String address) {
-        descriptor.setAddress(new Address(address));
-        return this;
-    }
+    // /**
+    //  * Sets the {@code Address} of the {@code EditExerciseDescriptor} that we are building.
+    //  */
+    // public EditExerciseDescriptorBuilder withAddress(String address) {
+    //     descriptor.setAddress(new Address(address));
+    //     return this;
+    // }
 
-    /**
-     * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code EditExerciseDescriptor}
-     * that we are building.
-     */
-    public EditExerciseDescriptorBuilder withTags(String... tags) {
-        Set<Tag> tagSet = Stream.of(tags).map(Tag::new).collect(Collectors.toSet());
-        descriptor.setTags(tagSet);
-        return this;
-    }
+    // /**
+    //  * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code EditExerciseDescriptor}
+    //  * that we are building.
+    //  */
+    // public EditExerciseDescriptorBuilder withTags(String... tags) {
+    //     Set<Tag> tagSet = Stream.of(tags).map(Tag::new).collect(Collectors.toSet());
+    //     descriptor.setTags(tagSet);
+    //     return this;
+    // }
 
-    public EditCommand.EditExerciseDescriptor build() {
-        return descriptor;
-    }
+    // public EditCommand.EditExerciseDescriptor build() {
+    //     return descriptor;
+    // }
 }
